refactor: created common settingspage component
